requirements.txt
transformers
pymongo
beautifulsoup4
requests
python-dotenv